fsmonitor--daemon: quit when the repository vanishes
Each worktree has a `.git` in its top-level directory, whether it is a
directory or a file. Once that directory or file is gone, it stops being
a Git worktree. So it is pointless to keep listening in that case.

This also helps the situation on Windows when a user wants to remove a
Git worktree altogether: a running daemon would prevent the deletion of
the worktree because its the latter is the former's working directory
(plus, it has to keep an open handle to that directory for that
`ReadDirectoryChangesW()` function to do its job).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>