fsmonitor: check fsmonitor valid flag when checking up to date
When using fsmonitor the CE_FSMONITOR_VALID flag should be checked when
wanting to know if the entry has been updated.  If the flag is set the
entry should be considered up to date and the same as if CE_UPTODATE is
set.

In order to trust the CE_FSMONITOR_VALID flag, the fsmonitor data needs to
be refreshed when the fsmonitor bitmap is applied to the index in
tweak_fsmonitor.  Also when merging trees in unpack_trees, if fsmonitor
was being used that should continue with the result index so that the
next git command will have the fsmonitor data to use instead of having to
check every entry in the index to see if it is valid.

Since the fsmonitor data is kept up to date for every command, some tests
needed to be updated to take that into account.