fsmonitor: do not check fsmonitor_dirty bits against cache entries on split index
The fsmonitor_dirty bits are set based on all the entries of both the index
and the base/shared index, where this check does not include the entries in
the base/shared index.  This means the fsmonitor_dirty will almost always
have more entries and hit this BUG.

At read time the fsmonitor_dirty bits are loaded and checked before the
base/shared index is read and entries from it are added to the cache.
During write the fsmonitor_dirty bitmap is built based on all the entries
of the in memory entries and not just the change index.

Since either making the fsmonitor_dirty bitmap to only have entries for
the one index or making sure when the check on the number of entries is
taking into account both indexes is a more extensive change, skip the
check on the number of entries when using a split index and fsmonitor.